+++ packages/devtools/devtools/src/models/settingsStore.ts
@@ -1,0 +1,36 @@
+import Store from '../utils/Store'
+
+const settingsStore = new Store({
+  showButton: true,
+  badgeType: 'collections',
+  trackQueries: true,
+  trackMutations: true,
+  trackMeasurements: true,
+})
+
+if (typeof document !== 'undefined') {
+  const initializeSettings = () => {
+    settingsStore.patch({
+      showButton: true,
+      badgeType: 'collections',
+      ...JSON.parse(localStorage.getItem('signaldb_devtools_settings') || '{}') as {
+        showButton?: boolean,
+        badgeType?: string,
+        trackQueries?: boolean,
+        trackMutations?: boolean,
+        trackMeasurements?: boolean,
+      },
+    })
+    settingsStore.subscribe(() => {
+      localStorage.setItem('signaldb_devtools_settings', JSON.stringify(settingsStore.get()))
+    })
+  }
+
+  if (['complete', 'interactive'].includes(document.readyState)) {
+    initializeSettings()
+  } else {
+    document.addEventListener('DOMContentLoaded', initializeSettings)
+  }
+}
+
+export default settingsStore